requirements.txt
aiokafka==0.5.2
aiomultiprocess==0.7.0
asyncpool==1.0
Cerberus==1.3.2
janus==0.5.0
kafka-python==1.4.6
pkg-resources==0.0.0
